move the egressFirewall_dns tests into the ovn package
Since the address_set is getting moved to a separate package there is
no longer a circular import with the address_set mock code and ovn
packages.
