+++ go-controller/pkg/ovn/egressfirewall_dns.go
@@ -181,16 +181,3 @@ func (e *EgressDNS) Shutdown() {
 func (e *EgressDNS) signalAdded(dnsName string) {
 	e.added <- dnsName
 }
-
-// function only used for testing
-// TODO when address_set gets moved to its own package remove this function and move
-// egress_firewall_dns_test into the same package as egressfirewall_dns
-func (e *EgressDNS) GetDNSEntry(dnsName string) (map[string]struct{}, []net.IP, addressset.AddressSet) {
-	e.lock.Lock()
-	defer e.lock.Unlock()
-	if dnsEntry, exists := e.dnsEntries[dnsName]; exists {
-		return dnsEntry.namespaces, dnsEntry.dnsResolves, dnsEntry.dnsAddressSet
-	}
-
-	return nil, nil, nil
-}
